requirements.txt
python-jose
requests
boto3
Jinja2